+++ demo/src/config/nodeLabels.js
@@ -1,0 +1,6 @@
+export default {
+  // Examples
+  'more-cowbell': '#0052CC',
+  'needs-review': '#FF8B00',
+  'research-has-failed-us': '#FF5630',
+};