+++ demo/src/plugins/dashboard/components/top-articles/TableRow.jsx
@@ -3,6 +3,7 @@ import React from 'react';
 import Collaborators from '@triniti/cms/plugins/raven/components/collaborators';
 import convertReadableTime from '@triniti/cms/utils/convertReadableTime';
 import Message from '@gdbots/pbj/Message';
+import NodeLabels from '@triniti/cms/plugins/ncr/components/node-labels';
 import NodeRef from '@gdbots/schemas/gdbots/ncr/NodeRef';
 import pbjUrl from '@gdbots/pbjx/pbjUrl';
 import UncontrolledTooltip from '@triniti/cms/plugins/common/components/uncontrolled-tooltip';
@@ -14,6 +15,7 @@ const TableRow = ({ idx, article }) => (
 
     <td>
       {article.get('title')}
+      <NodeLabels node={article} />
       <Collaborators nodeRef={NodeRef.fromNode(article)} />
     </td>